Cleanup Samples
* Move NHibernate initialization to Windsor installer;
* Move Windsor FilterAttribute installation to SharpArch.Web.Mvc.Castle
