+++ Samples/TardisBank/Solutions/Suteki.TardisBank.Web.Mvc/Suteki.TardisBank.Web.Mvc.csproj
@@ -256,6 +256,7 @@
     <Compile Include="CastleWindsor\HandlerInstaller.cs" />
     <Compile Include="CastleWindsor\MvcGooInstaller.cs" />
     <Compile Include="CastleWindsor\MvcServiceLocator.cs" />
+    <Compile Include="CastleWindsor\NHibernateInstaller.cs" />
     <Compile Include="CastleWindsor\QueriesInstaller.cs" />
     <Compile Include="CastleWindsor\RepositoriesInstaller.cs" />
     <Compile Include="CastleWindsor\ServicesInstaller.cs" />
